+++ Assets/Scenes/WorkInProgress/SzymanTestScene.unity
@@ -157,8 +157,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   speed: 10
   sensitivity: 90
-  playerBody: {fileID: 206836138}
-  controller: {fileID: 206836139}
 --- !u!23 &206836136
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -216,7 +214,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 206836134}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -3.97, y: 4.46, z: 0}
+  m_LocalPosition: {x: -11.53, y: 2.32, z: 8.78}
   m_LocalScale: {x: 2, y: 2, z: 2}
   m_Children:
   - {fileID: 1306738826}
@@ -268,7 +266,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 531005417}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -8.528, y: 0.1, z: 9.806}
+  m_LocalPosition: {x: -8.528, y: 0.09, z: 9.806}
   m_LocalScale: {x: 3.2756596, y: 0.99999976, z: 2.8891883}
   m_Children: []
   m_Father: {fileID: 1378206077}
@@ -379,7 +377,7 @@ GameObject:
   - component: {fileID: 1306738828}
   - component: {fileID: 1306738827}
   m_Layer: 0
-  m_Name: Main Camera
+  m_Name: MainCamera
   m_TagString: MainCamera
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -393,7 +391,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1306738825}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.109999895, y: 0.68999994, z: 0.36}
+  m_LocalPosition: {x: 0.109999895, y: 0.81, z: 0.31}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 206836138}
